python_sdk/pyproject.toml
[project]
name = "fastmcp-mre-breaking-task-group-with-invalid-arguments"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["fastmcp"]

[project.scripts]
fastmcp-mre-breaking-task-group-with-invalid-arguments = "simple_echo_mre:main"
